+++ src/Ext/Script/Body.h
@@ -13,6 +13,7 @@
 #include <WarheadTypeClass.h>
 #include <SpawnManagerClass.h>
 
+#include <TechnoClass.h>
 #include <Ext/House/Body.h>
 #include <Ext/Team/Body.h>
 #include <Utilities/Container.h>
@@ -90,6 +91,21 @@ enum class PhobosScripts : unsigned int
 	StopForceJumpCountdown = 16002,
 	RandomSkipNextAction = 16003,
 	PickRandomScript = 16004,
+	ConditionalJumpResetVariables = 16012,
+	ConditionalJumpManageResetIfJump = 16013,
+	AbortActionAfterSuccessKill = 16014,
+	ConditionalJumpManageKillsCounter = 16015,
+	ConditionalJumpSetCounter = 16016,
+	ConditionalJumpSetComparatorMode = 16017,
+	ConditionalJumpSetComparatorValue = 16018,
+	ConditionalJumpSetIndex = 16019,
+	ConditionalJumpIfFalse = 16020,
+	ConditionalJumpIfTrue = 16021,
+	ConditionalJumpKillEvaluation = 16022,
+	ConditionalJumpCheckCount = 16023,
+	ConditionalJumpCheckAliveHumans = 16024,
+	ConditionalJumpCheckObjects = 16025,
+	ConditionalJumpCheckHumanIsMostHated = 16026,
 
 	// Range 18000-18999 are variable actions
 	LocalVariableSet = 18000,
@@ -215,6 +231,23 @@ public:
 	static void PickRandomScript(TeamClass* pTeam, int idxScriptsList);
 	static void UnregisterGreatSuccess(TeamClass* pTeam);
 	static void SetCloseEnoughDistance(TeamClass* pTeam, double distance);
+
+	static void ConditionalJumpIfTrue(TeamClass* pTeam, int newScriptLine);
+	static void ConditionalJumpIfFalse(TeamClass* pTeam, int newScriptLine);
+	static void ConditionalJump_KillEvaluation(TeamClass* pTeam);
+	static void ConditionalJump_ManageKillsCounter(TeamClass* pTeam, int enable);
+	static void ConditionalJump_SetIndex(TeamClass* pTeam, int index);
+	static void ConditionalJump_SetComparatorValue(TeamClass* pTeam, int value);
+	static void ConditionalJump_SetComparatorMode(TeamClass* pTeam, int value);
+	static void ConditionalJump_SetCounter(TeamClass* pTeam, int value);
+	static void SetAbortActionAfterSuccessKill(TeamClass* pTeam, int enable);
+	static void ConditionalJump_ResetVariables(TeamClass* pTeam);
+	static void ConditionalJump_CheckHumanIsMostHated(TeamClass* pTeam);
+	static void ConditionalJump_CheckAliveHumans(TeamClass* pTeam, int mode);
+	static void ConditionalJump_CheckObjects(TeamClass* pTeam);
+	static void ConditionalJump_CheckCount(TeamClass* pTeam, int modifier);
+	static void ConditionalJump_ManageResetIfJump(TeamClass* pTeam, int enable);
+
 	static void SetMoveMissionEndMode(TeamClass* pTeam, int mode);
 	static void SkipNextAction(TeamClass* pTeam, int successPercentage);
 	static FootClass* FindTheTeamLeader(TeamClass* pTeam);
@@ -266,4 +299,5 @@ private:
 	static bool MoveMissionEndStatus(TeamClass* pTeam, TechnoClass* pFocus, FootClass* pLeader, int mode);
 	static void ChronoshiftTeamToTarget(TeamClass* pTeam, TechnoClass* pTeamLeader, AbstractClass* pTarget);
 	static void UpdateEnemyHouseIndex(HouseClass* pHouse);
+	static bool ConditionalJump_MakeEvaluation(int comparatorMode, int studiedValue, int comparatorValue);
 };