+++ src/Ext/Techno/Body.cpp
@@ -10,6 +10,7 @@
 #include <Ext/WarheadType/Body.h>
 #include <Ext/WeaponType/Body.h>
 #include <Ext/Script/Body.h>
+#include <Ext/Team/Body.h>
 
 #include <Utilities/AresFunctions.h>
 #include <Utilities/EnumFunctions.h>