+++ docs/Whats-New.md
@@ -283,6 +283,7 @@ New:
 - Repair Legless Cyborgs In Transports (by FS-21)
 - Override target under EMP attack behavior (By FS-21)
 - Mind Control Threshold (by FS-21)
+- Script actions for manipulating script flows with conditional jumps (by FS-21)
 - Customizable ROF random delay (by Starkku)
 - Animation with `Tiled=yes` now supports `CustomPalette` (by ststl)
 - New AI teams selector (by FS-21)